requirements.txt
# === Backend core ===
fastapi==0.115.0
uvicorn[standard]==0.30.1

# === Database ===
sqlalchemy==2.0.30
pymysql==1.1.1

# === Auth & security ===
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.1.1

# === Environment & uploads ===
python-dotenv==1.0.1
python-multipart==0.0.9


# ================================
# How to run the project
# ================================
# 1) Create and activate virtual environment:
#    python -m venv venv
#    source venv/bin/activate        (Linux/Mac)
#    venv\Scripts\activate           (Windows)
#
# 2) Install requirements:
#    pip install -r requirements.txt
#
# 3) Start backend server (from project root):
#    uvicorn backend.app.main:app --reload --env-file backend/.env --port 8000
#
# 4) Start frontend (from frontend/ folder):
#    npm install
#    npm run dev
#
# Backend runs on:   http://127.0.0.1:8000
# Frontend runs on:  http://127.0.0.1:5173 (default Vite port)


#pip install -r requirements.txt  run this in terminal so you have all the deps install
# this is for backend npm install && npm run dev

